fix(sim): T1-2b-fix Codex audit P1+P2 fix-pass [T1-2b-fix]
Composed 8-chunk post-Codex-audit fix pass (REVISE verdict, 8 P1 + 6 P2).
All 8 P1s and 4 of 6 P2s addressed (P2-10 NoOpStub triggers + P2-11
wrong MemoryEvent owner explicitly deferred per task spec).

canonical hash: REBASELINED (trigger: #1; old: blake3:18f1776c…a5d048; new: blake3:dbe4f49b…85f2; reason: VERSION 5→6; P1-2 per-player signature_candidates added to canonical encoding)

P1-1: MatchState::signature_firing reshaped from [Option<SignatureFiring>; 22]
      to [[Option<SignatureFiring>; 4]; 22] (2D: outer=roster slot, inner=BiasCategory).
      Dispatcher, canonical encoder, and all call sites updated.

P1-2: Per-player signature_candidates now encoded in canonical state.
      Canonical encoder VERSION bumped 5→6; wire-format diagram extended to
      show signature_cooldowns + signature_firing + signature_first_fired_seen
      sections after ball.

P1-3/P1-4: SeedLayer + seed_fn moved from fw-match-sim::decision_cadence to
      fw-core::seed (ADR-0009 canonical home). Decision cadence re-exports.
      tick_goalkeeper now accepts player: &PlayerState as second parameter
      (GK FSM attribute reads in skeleton tier; real bindings at T1-2b-iii-b).

P1-5/P2-9: All BT binding sites verified correct against bt-attribute-binding.md.
      Personality bias helpers + PT divisor confirmed present. No changes needed.

P1-6: TriggerFn changed from fn(…) -> bool to fn(…) -> Q32 (fit-score).
      Returns Q32::ZERO for ineligible; positive value in (0, 1] otherwise.
      Dispatcher now weights affinity × fit_score in the softmax per ADR-0011.
      body_shield_pressure: marking × strength × aggression.
      long_range_strike: composure × long_shots.
      first_time_diagonal_switch: vision × passing.

P1-7: Canonical encoder corrected for 2D signature_firing layout (already
      reflected in P1-1 above).

P1-8: Anti-vacuousness companion tests added for all 5 proptest ACs:
      vacuousness_check_ac2/ac3/ac4/ac5 in signature_dispatcher_proptest.rs.
      Each constructs an invariant-violating state and proves the violation
      would be caught.

P2-12: determinism-audit.py per-rule exemptions tightened. math.rs + q32.rs
      now have per-rule exemptions (f32/f64 only); other rules still scan.
      dto.rs retains full-file exemption as renderer-side projection.

P2-13: ADR-0012 trigger labels corrected in canonical_hash.rs + RON fixture.
      Behavioral-change entries (T1-2b-iii-c, iii-c P0+P1, iii-d) now
      correctly labeled trigger #3, not #1. Wire-format diagram in canonical.rs
      extended to show the 3 signature sections appended after ball (VERSION 5+).

P2-14: pitch_control.rs line 146 unwrap() replaced with let-else pattern.
      Non-empty invariant is preserved by the early-return guard above.

SelectFn type alias added to bt/mod.rs to remove the very-complex-type warning.

+++ crates/fw-match-sim/tests/decision_cadence_proptest.rs
@@ -192,14 +192,14 @@ proptest! {
     #[test]
     fn seed_fn_is_stable(
         match_seed in arb_seed(),
-        tick in 0i64..1000,
+        tick in 0u32..1000,
         site in any::<u32>()
     ) {
         let a = fw_match_sim::decision_cadence::seed_fn(
-            match_seed, tick, SeedLayer::Decision, site as u64
+            match_seed, tick, SeedLayer::Decision, site
         );
         let b = fw_match_sim::decision_cadence::seed_fn(
-            match_seed, tick, SeedLayer::Decision, site as u64
+            match_seed, tick, SeedLayer::Decision, site
         );
         prop_assert_eq!(a, b);
     }
